Moved common parts of E2E tests into ConfigTestTemplate.
